+++ src/constructs/iam/roles/index.ts
@@ -1,4 +1,3 @@
-export * from "./fastly-kinesis-log";
 export * from "./github-actions";
 export * from "./instance-role";
 export * from "./roles";
