+++ src/experimental/constructs/iam/roles/fastly-kinesis-log.test.ts
@@ -1,21 +1,27 @@
 import { Template } from "aws-cdk-lib/assertions";
-import { FASTLY_AWS_ACCOUNT_ID } from "../../../constants/fastly-aws-account-id";
-import { simpleGuStackForTesting } from "../../../utils/test";
-import { GuKinesisStream } from "../../kinesis";
-import { GuFastlyKinesisLogRole } from "./";
+import { FASTLY_AWS_ACCOUNT_ID } from "../../../../constants";
+import { GuKinesisStream } from "../../../../constructs/kinesis";
+import { simpleGuStackForTesting } from "../../../../utils/test";
+import { GuFastlyKinesisLogRoleExperimental } from "./fastly-kinesis-log";
 
 describe("The GuFastlyKinesisLogRole construct", () => {
   it("correctly wires up the policy", () => {
     const stack = simpleGuStackForTesting();
     const testStream = new GuKinesisStream(stack, "testStream");
-    new GuFastlyKinesisLogRole(stack, "testKinesisLogRole", { stream: testStream, roleName: "writeToKinesisRoleTest" });
+    new GuFastlyKinesisLogRoleExperimental(stack, "testKinesisLogRole", {
+      stream: testStream,
+      roleName: "writeToKinesisRoleTest",
+    });
     expect(Template.fromStack(stack).toJSON()).toMatchSnapshot();
   });
 
   it("assumes the correct role", () => {
     const stack = simpleGuStackForTesting();
     const testStream = new GuKinesisStream(stack, "testStream");
-    new GuFastlyKinesisLogRole(stack, "testKinesisLogRole", { stream: testStream, roleName: "writeToKinesisRoleTest" });
+    new GuFastlyKinesisLogRoleExperimental(stack, "testKinesisLogRole", {
+      stream: testStream,
+      roleName: "writeToKinesisRoleTest",
+    });
 
     Template.fromStack(stack).hasResourceProperties("AWS::IAM::Role", {
       AssumeRolePolicyDocument: {
