+++ src/experimental/constructs/iam/roles/fastly-kinesis-log.ts
@@ -1,10 +1,10 @@
 import { AccountPrincipal } from "aws-cdk-lib/aws-iam";
-import { FASTLY_AWS_ACCOUNT_ID } from "../../../constants/fastly-aws-account-id";
-import type { GuStack } from "../../core";
-import { GuFastlyCustomerIdParameter } from "../../core";
-import type { GuKinesisStream } from "../../kinesis";
-import { GuKinesisPutRecordsPolicy } from "../policies/kinesis-put-records";
-import { GuRole } from "./roles";
+import { FASTLY_AWS_ACCOUNT_ID } from "../../../../constants/fastly-aws-account-id";
+import type { GuStack } from "../../../../constructs/core";
+import { GuFastlyCustomerIdParameter } from "../../../../constructs/core";
+import { GuRole } from "../../../../constructs/iam";
+import type { GuKinesisStream } from "../../../../constructs/kinesis";
+import { GuKinesisPutRecordsPolicyExperimental } from "../../policies/kinesis-put-records";
 
 export interface GuFastlyKinesisLogRoleProps {
   /**
@@ -26,18 +26,19 @@ export interface GuFastlyKinesisLogRoleProps {
  * the Guardian's Fastly customer id.
  *
  */
-export class GuFastlyKinesisLogRole extends GuRole {
+export class GuFastlyKinesisLogRoleExperimental extends GuRole {
   constructor(scope: GuStack, id: string, props: GuFastlyKinesisLogRoleProps) {
     const fastlyCustomerId = GuFastlyCustomerIdParameter.getInstance(scope).valueAsString;
+    const { roleName, stream } = props;
 
     super(scope, id, {
-      roleName: props.roleName,
+      roleName,
       assumedBy: new AccountPrincipal(FASTLY_AWS_ACCOUNT_ID),
       externalIds: [fastlyCustomerId],
     });
 
-    const policy = new GuKinesisPutRecordsPolicy(scope, "GuKinesisPutRecordsPolicy", {
-      stream: props.stream,
+    const policy = new GuKinesisPutRecordsPolicyExperimental(scope, "GuKinesisPutRecordsPolicy", {
+      stream,
     });
 
     policy.attachToRole(this);
