+++ src/experimental/constructs/iam/roles/index.ts
@@ -1,0 +1,1 @@
+export * from "./fastly-kinesis-log";
