+++ src/experimental/constructs/policies/index.ts
@@ -1,0 +1,1 @@
+export * from "./kinesis-put-records";
